+++ src/main/java/com/memo/memo/dto/MemoDto.java
@@ -7,7 +7,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 @Schema(description = "메모 정보를 담는 DTO")
 public class MemoDto {
     @Schema(description = "메모 ID", example = "1", defaultValue = "1")
-    private Long id;
+    private Integer id;
 
     @Schema(description = "메모 내용", example = "Memo Swagger Test!" , allowableValues = {"메모1", "메모2"})
     private String content;
@@ -15,16 +15,16 @@ public class MemoDto {
     public MemoDto() {
     }
 
-    public MemoDto(Long id, String content) {
+    public MemoDto(Integer id, String content) {
         this.id = id;
         this.content = content;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
